Create diagnotician by superadmin completed
