+++ src/main/java/org/automation/config/config.properties
@@ -1,4 +1,4 @@
 browser=chrome
 url=https://topuptalent.com/Diagnosticlearning/
-userName=shawn
+userName=tim
 password=123456
