+++ testng.xml
@@ -4,7 +4,10 @@
     <test thread-count="1" parallel="classes" name="Test">
 
         <packages>
-            <package name="test.*"></package>
+<!--            <package name="test.*"></package>-->
         </packages>
+        <classes>
+            <class name="test.SuperAdminTest"></class>
+        </classes>
     </test>
 </suite>
